trello-sync/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic-settings==2.5.0
